Fix for Manifest.plist parsing on Windows.
Due to a bug in the `Mac::PropertyList` Perl module, binary plist
files (Manifest.plist, Status.plist) are sometimes randomly
appear to be corrupted when running on Windows. When this happens,
the script usually ends with the infamous `Not a HASH reference
at ios_backup_extractor.pl line 729` error.

Before this bug is fixed in an upstream module [1], this change
explicitly forces Manifest.plist and Status.plist to be read in
a binary mode, thus avoiding the problematic code path.

[1]: https://github.com/briandfoy/mac-propertylist/pull/26